+++ include/roboteam_ai/stp/invariants/WeHaveBallInvariant.h
@@ -5,12 +5,12 @@
 #ifndef RTT_WEHAVEBALLINVARIANT_H
 #define RTT_WEHAVEBALLINVARIANT_H
 
-#include "stp/invariants/BaseInvariant.h"
+#include "BaseInvariant.h"
 
 namespace rtt::ai::stp::invariant {
 class WeHaveBallInvariant : public BaseInvariant {
    public:
-    [[nodiscard]] double metricCheck(world_new::view::WorldDataView world, const Field *field) const noexcept override;
+    [[nodiscard]] uint8_t metricCheck(world_new::view::WorldDataView world, const Field *field) const noexcept override;
 };
 }  // namespace rtt::ai::stp::invariant
